@@ -1,21 +1,24 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFramework>netcoreapp2.1</TargetFramework>
+		<TargetFramework>net5.0</TargetFramework>
 		<RuntimeFrameworkVersion>2.1.0</RuntimeFrameworkVersion>
 		<PreserveCompilationContext>true</PreserveCompilationContext>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="FluentAssertions" Version="4.19.4" />
-		<PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="5.0.1" />
-		<PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.1.0" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="2.1.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.1.0" />
-		<PackageReference Include="Moq" Version="4.8.3" />
-		<PackageReference Include="StructureMap" Version="4.6.1" />
-		<PackageReference Include="StructureMap.Microsoft.DependencyInjection" Version="1.4.0" />
-		<PackageReference Include="xunit" Version="2.3.1" />
-		<PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
+		<PackageReference Include="FluentAssertions" Version="5.10.3" />
+		<PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="9.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="5.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="5.0.0" />
+		<PackageReference Include="Moq" Version="4.15.1" />
+		<PackageReference Include="StructureMap" Version="4.7.1" />
+		<PackageReference Include="StructureMap.Microsoft.DependencyInjection" Version="2.0.0" />
+		<PackageReference Include="xunit" Version="2.4.1" />
+		<PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
+		  <PrivateAssets>all</PrivateAssets>
+		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
 		<DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
 	</ItemGroup>
 
