@@ -1,7 +1,7 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
 	<PreserveCompilationContext>true</PreserveCompilationContext>
   </PropertyGroup>
 
@@ -51,13 +51,17 @@
 
   <ItemGroup>
     <PackageReference Include="CPermissions" Version="1.1.0" />
-    <PackageReference Include="filer.entityframeworkcore" Version="0.0.8-beta2" />
-    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="5.0.1" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.1.0" />
-    <PackageReference Include="Sendgrid" Version="9.9.0" />
-    <PackageReference Include="structuremap.aspnetcore" Version="1.4.0" />
-    <PackageReference Include="Z.EntityFramework.Plus.EFCore" Version="1.7.20" />
+    <PackageReference Include="Filer.EntityFrameworkCore" Version="0.0.11-beta" />
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="9.0.0" />
+    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.16.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="5.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="5.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.0" />
+    <PackageReference Include="Sendgrid" Version="9.21.2" />
+    <PackageReference Include="structuremap.aspnetcore" Version="2.0.0" />
+    <PackageReference Include="Z.EntityFramework.Plus.EFCore" Version="5.1.2" />
   </ItemGroup>
 
   <ItemGroup>
